+++ src/index.ts
@@ -1,2 +1,2 @@
-export { CustomQueryBuilder, CustomQueryBuilderError, defineQueryBuilder } from './CustomQueryBuilder';
-export type { LeftJoinsAndSelects, JoinsAndSelects, QueryBuilderExtensions } from './CustomQueryBuilder';
+export { CustomQueryBuilder, CustomQueryBuilderError, defineQueryBuilder, defineSharedQueryBuilder } from './CustomQueryBuilder';
+export type { LeftJoinsAndSelects, JoinsAndSelects, QueryBuilderExtensions, SharedQueryBuilder } from './CustomQueryBuilder';